tg-10288 Modificar flujo de descarga para recetas

+++ back-end/hospital-api/src/main/java/net/pladema/clinichistory/requests/medicationrequests/service/CreateMedicationRequestService.java
@@ -4,5 +4,5 @@ import net.pladema.clinichistory.requests.medicationrequests.service.domain.Medi
 
 public interface CreateMedicationRequestService {
 
-    Integer execute(MedicationRequestBo medicationRequest);
+    Long execute(MedicationRequestBo medicationRequest);
 }
